oneplus-oneplus3: workaround bug with touchscreen
libinput, which I believe, claims there is a kernel bug with the
synaptics driver touchscreen in use.

This is the recommended fix according to PostmarketOS.

 * https://wiki.postmarketos.org/wiki/Troubleshooting:touchscreen#sec_touchscreen:_kernel_bug

I have not seen any issues from using this fix.

The error:

```
synaptics: kernel bug: device has min == max on ABS_MT_WIDTH_MAJOR
```

According to the touchscreen documentation of the kernel, I don't think
that it is an issue to remove this information from touch events.