Remove postmarketOS-devices.json
Anchoring ourselves to an external project, at this point, is not
feasible. This does not allow us to explore making better use of the
capabilities of nix.

Though, a mechanism using third-party resources, like the postmarketOS
device information, is something that is desirable. The difference is
that it would need to *import* the information in the format used by
mobile-nixos, rather than mobile-nixos importing the information from a
"data dump".