google-walleye: Inits kernel at 4.4.195
The kernel build is technically the same as google-taimen, with some
options switched around. An upcoming improvement will be to share the
"wahoo" platform's kernel build with an option to switch between the
two at build time, and further along the line, with modular kernels,
sharing the same exact builds, but with the device-specifics as modules.