rootfs: handle cross-compilation with a dummy rootfs
Note that this rootfs *will not boot*, and it's by design.

It does not have the appropriate filesystem label, nor does it have the
right UUID.

This may look absurdly useless, but I assure you it's not.

With this scheme, some targets are easier to deal with. Namely, those
that allow booting from multiple boot sources. With the u-boot system
type, you can produce a full disk image that can be burned on an SD
card, and will boot the internal storage's rootfs that was already
present.

This makes it much more trivial to deal with stage-1 development while
sitll booting the internal rootfs.

Though, I'm not expecting this behaviour to *remain*.

(cherry picked from commit f8dbc5afa6beb91b6c1576ff02767ede7c1a267e)