motorola-addison: Remove `console=` argument
Ah! It looks like this is not a constant, but a common failure mode
among android-based devices!

In dmesg:

```
Warning: unable to open an initial console.
```

Simply removing the `console=` fixes the console.

See f295dc1739d1776025a1c21e0e3a9e8780a4062a for a similar instance of
the issue.