libhybris: Allow relying on legacy properties
This is used to get a slimmer libhybris for adbd for stage-1.

The previous behaviour (before 35976c201741cfc21c114615b8524db5367e034a)
was to rely on a "homegrown" system to get properties.

Starting with that new properties system, it will rely on the vendor
libraries, thus requiring all of them, plus the linker, to get
properties.

Turns out, adbd doesn't actively need anything from the vendor libraries
or from the actual propeties system.

So, it's fine to revert the change to get a working libhybris for
stage-1.

Though, what would be even better is to strip out only the minimum
requirements for adbd, and make a "mini hybris" that we can rely on for
adbd.