kernel-builder: Fix menuconfig for more recent kernels
While staying compatible with older ones.

These changes are required since recent kernels' nconfig expect that the
Makefile will have setup the environment for *many* things.

We can't simply run nconfig directly anymore.

What's that about `run-nconfig`?

Well, `make nconfig` tries to build again, in a situation where it won't
be able to build. *sigh*. With this, we take the assumedly-fine-to-run
nconfig that is already in our botched up source tree.