nixos-integration: Use initrd.enable = false
That new option allows us to remove the crutch of pretending to be a
container.

Pretending to be a container had a couple drawbacks, they are now
irrelevant.

The dummy outputs could be re-worked to instead use out own stage-1
infra, but I prefer using dummies for now, for better insulation.