system/android: Review recovery build
Instead of borrowing some of its system.build products, rely instead on
a full system.build for the recovery.img.

This allows us to better make use of the implicit declarativeness of the
system configuration. At some point the recovery "system" will be moved
out of that file and things will continue to work.