treewide: Remove installer/mobile-installer dead weight
This didn't end up being something that's being worked on. It probably
was entirely confusing for everyone as it wasn't documented, and didn't
actually do anything useful most of the time.

Rather than polluting the build with some one-off special cases like
this, we'll implement an "installer" as an example system, if we even
want an installer.

It's not planned yet *how* users should install, it may not even be an
installer type system.

This also rewords the "successful default build failure" message.