qualcomm: Add fb notify quirk
This quirk touches the `blank` file of the framebuffer /sys nodes, which
ends up unsuspending things that start suspended until the *right thing*
is happening to not be suspended.

X11, among others, do the right things. It seems other framebuffer
interfaces are not.

This can likely be fixed in other ways, but this is the more
approachable way to me right now.