potter: invert card detect gpio sense for mmc1
This is needed to make mobile-nixos notice that an external SD card is
present.  Looking at the device tree sources for various msm8953
devices, it seems to be pretty much a 50/50 chance whether the card
detect pin is active high or low.  Of course that *might*
just mean it's broken on half of all phones

See also https://lore.kernel.org/patchwork/patch/615679/ - clearly I
amn not the only person who finds this confusing