kernel-builder: Implement kernel normalization as a nix build
This uses the actual output expected by the build, rather than hoping
menuconfig gives the same result.

For kernels prior to 5.8 it wasn't an issue.

Starting with 5.8 the build environment will influence the actual
configuration used.

We're stripping some information when comparing.

The exact TEXT_VERSION string will change according to whether it's a
native or cross compilation.

Finally, minor version bumps are not relevant here.