kernel-builder: Work around kernels getting compiled two times
It looks like this is not a regression from the latest changes to the
kernel builder. Testing on top of ffdba8fecbb2aee6fb79e4a1cde5e06035ef407b
exhibits the same issue.

This change works around the issue by doing it with one invocation of
`make`, rather than doing it twice.

It has not been figured out at which exact version of the kernel this
stops being an issue, and if it is an issue from mainline Linux
outright, or something that came from Android-specific modifications.

Though it has been verified that for 3.18, both Qualcomm and Mediatek
source trees exhibit the same issue.

Fixes #216