xiaomi-begonia: Don't rely on bundled dtc_overlay
While discussing with other devs, working on a samsung device on another
distro, I figured out it was basically free to try and build it with a
shim that's calling "proper" `dtc` and see if it works.

It worked for xiaomi-begonia.

It worked for that other device (on postmarketOS) too!

This means we're suddently gaining the ability to build natively.
Additionally, for postmarketOS, it allows them to build, considering
`dtc_overlay` uses glibc, and postmarketOS doesn't.