devices/uefi-x86_64: Init, from qemu-x86_64
For the time being, this is mostly a copy of the QEMU system.

The QEMU system is slated to be removed, once the UEFI system provides
the necessary tooling to get a VM going easily.

We'll probably want to add other generically-useful modules to that
system.

We'll have to look at NixOS's boot.initrd.availableKernelModules and
boot.initrd.kernelModules options.