kernel-builder: Call either of install or zinstall
It seems that under specific conditions:

 - 2a5 hardware
 - -j22

the call to `make install zinstall` (simplified here) may break, with
the installation of `System.map` happening in a manner where it fails
the build. It was not trivial to reproduce elsewhere than on a 2a5
machine, for unknown reasons.

It was also only observed on google-walleye, but there is no reason is
shouldn't be failing on other devices too.

Odd.