fix build failure by providing kernel filename to internal NixOS option
The update from 29a10085f313904516992684f1d0fd647fb2ad6d caused the
uncompressed kernel to not be installed anymore into the kernel output.

This meant that building the top-level argument failed as a file assumed
to exist wasn't there anymore.

We now provide the appropriate filename when we know it.