STATUS: WIP, read more...
It currently requires Tianocore for i686, which does not build. I was
testing against an arbitrary binary build of it from another distro.

Alternatively, u-boot for i686 could be another way to get this going.

**But why?**

This allows us to test *some* parts of Mobile NixOS against a 32 bit
target. Except that it is not equivalent enough with armv7 to be worth
actually spending a lot of time working on.