bin/hydra-eval: Add hydra-like eval helper
This helper is intended to be used by contributors that are changing
parts of this project that are either touching `lib/` and the evaluation
helpers, or changing `release.nix`.

There are no facilities to *compare* evals, yet. But this can be done by
the contributor, they can go to any commit they want to compare to, run
the eval and save the output markdown report. Then they can compare
against their new markdown report.

Finally, splitting the eval errors like we do in the reports should make
it really obvious what errors happened during the eval.