mruby: use the right rake to build
The wrapper for bundlerApp, starting with
e8993319724901f132dd4ed997f5383c4a482940 in Nixpkgs ends up using the
runtimeShell for the target platform. Yes, with buildPackages.

This forces an mruby that is more appropriate, rather than rely on
buildPackages.

It may look like it works for aarch64 without this patch, but this is
only because for pkgsStatic+aarch64 cross-compiles fine.