potter: coalesce seven DTBs into three for space saving
Moto Potter boot partition is only 16MB, which is tight. In the vendor
kernel it has seven hwrev variants with a dts file for each - but four
of them are identical to three of the others. We can make a
significant saving on image size by shipping one of each of the
differing dts files and adding the board-id values of the other
identical hwrevs to each.

The new files live in devices/motorola-potter/kernel/dts and are
copied into the kernel source tree at build time. It may be that there
is a better way to do this: if not, it may be that there should be a
more general mechanism for copying files into the kernel tree that's
not limited to arch/arm/boot/dts, but that's out of scope for this
commit