potter: coalesce seven DTBs into three for space saving
Moto Potter boot partition is only 16MB, which is tight. In the vendor
kernel it has seven hwrev variants with a dts file for each - but four
of them are identical to three of the others. We can make a
significant saving on image size by shipping one of each of the
differing dts files and adding the board-id values of the other
identical hwrevs to each.