fix panel off-by-5000 bug
CONFIG_SCREEN_OFF_GESTURES adds 5000 to the reported X co-ordinate
when it believes the screen is suspended, but it always believes the
screen is suspended even when it isn't. Hypothesis is that (1)
the pm_ops are not supplied in the i2c_driver code, so the
`suspended=false` line in synaptics_rmi4_resume is not called by
that path; (2) something odd is happening in synaptics_dsx_panel_cb
- which is passed to fb_register_client - such that the equivalent
message from the framebuffer is not being recognised as a reason to
call synaptics_rmi4_resume then either. </Handwaving>