fix touchscreen without need of quirk
mdss_fb_release_all runs when the display blanks between the splash
screen and launching the main menu (in "hello" example). If ref_cnt is
zero, it calls fb_notifier_call_chain with the payload of
FB_BLANK_POWERDOWN. The synaptics_dsx_i2c touchscreen driver is
subscribed to this notifier, and disables touch.

This is all fine except that the corresponding opposite operation - to
notify the subscribers with FB_BLANK_UNBLANK when the display
is *opened* - is not done anywhere. Add a call to mdss_fb_open
to do this whenever it's calling mdss_fb_blank_sub

For context, this behaviour was introduced in

   commit a12be25aec3fd648f2a9c4c27e88d470bed41f67
   Author: parasur <parasur@motorola.com>
   Date:   Mon Nov 7 09:45:54 2016 +0800
   https://github.com/LineageOS/android_kernel_motorola_msm8953/commit/a12be25aec3fd648f2a9c4c27e88d470bed41f67#diff-eab94fa9ea3e15183d23b93fbaf4e08bc137ae4ebbdc589fc106355b753d1b6a

but probably (speculating) goes unnoticed in Androd because the
framebuffer is never closed in Android unless something crashes, at
which point most people probably reboot anyway