kernel-builder: Validate structured kernel configuration
While currently this is not plugged into anything, the intent is to
provide the structured configuration from the modules system.

Different modules will be able to provide kernel configuration options
that are required (or suggested), and the validation will fail the build
if the options are not provided.

A further update will add those options for normalization, allowing the
update to be automatic.

This, finally, is the first move towards providing an alternative route
for kernel configuration; "defconfig + structured config", where a
platform-provided defconfig can be used as a base, and the structured
config "finishes" configuration. Hopefully this should allow using the
defconfig files from the vendor repositories directly, obviating the
need to *somehow* provide a full configuration file.

Note that the full configuration file scheme is expected to stay.

Huh, where's that `eval-config` coming from? ██████████████████
█████ ████████████████ ██████████. Coming soon!