[HACK] Add `outputs.u-boot.temp-tow-boot-install-script`
This is not meant to be merged.

This is meant as a way forward for dogfooding the new Tow-Boot setup
before having the installer system images working.

While it won't be merged, it is recommended users start setting-up their
systems using this script as it will be equivalent to what the final
installed systems will look like.