system-types/u-boot: Drop U-Boot handling
This is because Mobile NixOS is taking a stance. Distros have already
enough to handle, handling what is basically a computer's BIOS is too
much of a burden, and makes things way more awkward than they should be.

By relying on a third-party project, which in turn relies on
standards-based boot, we can start building toward a better future not
only for Mobile NixOS, but for all Linux on ARM.