motorola-potter: add gadgetfs support for adb in stage-2
the problem we are trying to solve is that when adbd is started in
stage-1, it does not survive to stage-2. This is due to two things

- the config in configfs becomes disabled ("UDC" file becomes empty)
- adbd process is killed

This is more a workaround than a resolution: perhaps the better
approach would be for stage-1 to clean up its configfs properly, and
then write a NixOS module for adbd, which would be responsible for
setting up its own config entries. Then it wouldn't be coupled to the
stage-1 settings.