examples/installer: Write the configuration file directly from config
This reduces some parts of the complexity. No need to re-interpret the
configuration in the scripted installer. The scripted installer *just*
has to re-use the existing configuration files.

The JSON file is still saved, even if redundant, I'm not sure how I want
to approach the partitioning/filesystems step yet.