examples/installer: Added disk formatter script
This formats the target device in a **totally opinionated** manner.

This was verified to work and format, but not verified to integrate well
in the installer *yet*.

It is assumed some "higher level" tool will provide the paths to the
installer script. Also this tool will know about the intrinsic
"internal" storage devices of the target devices.