examples/installer: Add pstore partition
Not used yet, but see:

 - https://docs.kernel.org/admin-guide/pstore-blk.html

With proper care, and with UMS mode in the platform firmware, this
should help users diagnose problems after the fact.

The partition is likely way too big. But I wanted to "round-up" to 32MiB
the whole mess of reserved partitions. We might end-up cutting into the
pstore partition if we ever want to add some necessary feature.

Note that while the partition will exist, end-users will be in control
of the pstore feature. It will be enabled by default, but it will be
made trivial to disable.