examples/installer: Add automated installer
This uses the configuration file we generated from the app, copies it to
the already mounted root mount point, then builds to to the target Nix
store, *then* activates the profile, and finally installs the boot
partition (which holds the stage-1 image).

All the logic to determine the boot partition target is held in the main
app, and this is also how the disk formatting script does it.

This way we have one source of truth to determine the target block
device: the graphical app, which coordinates the whole setup.