examples/installer: Revise how phone environment config is handled
The goal here is to ensure the configuration used will not deviate from
the examples system. This, in turn, ensures there is cache coverage for
the given Nixpkgs/Mobile NixOS eval pair.

In other word, if the `tested` jobset passes, we know the installer will
be able to use the cache.