kernel/builder: Review approach for menuconfig pkg-config use
This is less brittle than the previous way. The previous way tried to
massage the existing script.

Let's instead replace the script outright with one that provides the
correct results.

The only way this will break, would have broken the previous way, is if
they decide to depend on more libraries or more output.