modules/stage-0: Ensure enabling/disabling is a no-op
This ensures that, for implementation details, bits that use stage-0 as
a synonym for stage-1 when it is not supported gets the same exact eval
than the one for stage-1.

Without this change, some `fdt-forward` and changes made under that
`mkIf` would end-up included. `fdt-forward` was found to be in the
stage-1 image even though it was irrelevant.