kernel: Add useStrictKernelConfig disabled by default
This changes the behaviour we've been relying on, by default, as it will
not fail when the provided config differs from the generated config.

This will help end-users trivially follow minor updates, and maybe even
major updates.

This does not affect normalization as those values are already set when
validating or generating the config.