support/kernel-config: disable ARM64_SME
This was marked as broken somewhere between 6.11.5 and 6.11.8 judging by me trying to update the PinePhone kernel.
It is also marked as broken in newer versions.
From https://lore.kernel.org/linux-arm-kernel/173097843612.164342.13696404397428904701.b4-ty@kernel.org/T/: "For now, mark SME as BROKEN in the hope that we can fix things properly in the near future. As SME is an OPTIONAL part of ARMv9.2+, and there is very little extant hardware, this should not adversely affect the vast majority of users."
Considering this, I think it's best to just completely disable this feature for now.