release.nix: Change semantics to require `pkgs` for input, and rely on pinned Nixpkgs otherwise
The only reason it was using `<nixpkgs>` was because it was how it got
its Nixpkgs within Hydra evals.

We're not on the nixos.org hydra anymore, so we can finally take it as
an input.

Follow the conventions set by `default.nix` in how `system` must be
provided only on `pkgs` if `pkgs` is provided, and otherwise rely on
`builtins.currentSystem` by default.